d8d92656 Update game settings and add user selections to console page
