+++ go-server/go.mod
@@ -1,14 +1,13 @@
-module liarOfTuring
+module liar-of-turing
 
 go 1.21.4
 
 require (
-	github.com/bmizerany/pat v0.0.0-20210406213842-e4b6760bdd6f
 	github.com/gorilla/websocket v1.5.1
 	github.com/rs/cors v1.10.1
 )
 
 require (
-	github.com/joho/godotenv v1.5.1 // indirect
+	github.com/joho/godotenv v1.5.1
 	golang.org/x/net v0.17.0 // indirect
 )
