+++ go-server/utils/utils.go
@@ -1,0 +1,14 @@
+package utils
+
+import "time"
+
+func GetCurrentTimestamp() int64 {
+	return time.Now().UnixNano() / int64(time.Millisecond)
+}
+
+func Max(a, b int) int {
+	if a > b {
+		return a
+	}
+	return b
+}
