+++ web-application/src/store/gameStore.ts
@@ -2,5 +2,5 @@ export const TIMER_TIME = 50;
 
 export const MAX_PLAYER = 6;
 
-export const GAME_ROUND_NUM = 3;
+export const GAME_ROUND_NUM = 1;
 export const GAME_TURNS_NUM = 2;
